Day-16 - Spatie permission setup and authorization finishes

+++ app/Http/Controllers/Backend/CategoryController.php
@@ -8,6 +8,7 @@ use App\Http\Requests\CategoryCreateRequest;
 use App\Models\Category;
 use App\Repositories\CategoryRepository;
 use Illuminate\Contracts\Support\Renderable;
+use Illuminate\Support\Facades\Auth;
 
 class CategoryController extends Controller
 {
@@ -23,6 +24,10 @@ class CategoryController extends Controller
      */
     public function index(CategoryDataTable $dataTable)
     {
+        if (!Auth::user()->hasPermissionTo('category.view')) {
+            abort(403, 'You are not authorized to view category');
+        }
+
         return $dataTable->render('backend.pages.category.index');
     }
 
@@ -33,6 +38,10 @@ class CategoryController extends Controller
      */
     public function create(): Renderable
     {
+        if (!Auth::user()->hasPermissionTo('category.create')) {
+            abort(403, 'You are not authorized to create category');
+        }
+
         return view('backend.pages.category.create', [
             'parentCategories' => $this->categoryRepository->printCategory()
         ]);
@@ -46,6 +55,10 @@ class CategoryController extends Controller
      */
     public function store(CategoryCreateRequest $request)
     {
+        if (!Auth::user()->hasPermissionTo('category.create')) {
+            abort(403, 'You are not authorized to create category');
+        }
+
         $category = $this->categoryRepository->create($request->all());
 
         if (!empty($category)) {
@@ -64,6 +77,10 @@ class CategoryController extends Controller
      */
     public function show(int $id)
     {
+        if (!Auth::user()->hasPermissionTo('category.view')) {
+            abort(403, 'You are not authorized to view category');
+        }
+
         $category = $this->categoryRepository->show($id);
     }
 
@@ -75,6 +92,10 @@ class CategoryController extends Controller
      */
     public function edit(Category $category): Renderable
     {
+        if (!Auth::user()->hasPermissionTo('category.edit')) {
+            abort(403, 'You are not authorized to edit category');
+        }
+
         return view('backend.pages.category.edit', [
             'category' => $category,
             'parentCategories' => $this->categoryRepository->printCategory($category->parent_id)
@@ -90,6 +111,10 @@ class CategoryController extends Controller
      */
     public function update(CategoryCreateRequest $request, Category $category)
     {
+        if (!Auth::user()->hasPermissionTo('category.edit')) {
+            abort(403, 'You are not authorized to edit category');
+        }
+
         $categoryUpdated = $this->categoryRepository->update($request->except('_token', '_method'), $category->id);
 
         if (!empty($categoryUpdated)) {
@@ -108,6 +133,10 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
+        if (!Auth::user()->hasPermissionTo('category.delete')) {
+            abort(403, 'You are not authorized to delete category');
+        }
+
         if ($this->categoryRepository->delete($category)) {
             session()->flash('success', 'Category deleted successfully.');
             return redirect()->route('admin.categories.index');
